Optional border added to png
When generating the png using convert, the trim is often too tight and things like block matrices are right at the edge and difficult to see.  Adding a border allows for a buffer to help viewing the previews.  This is a simple version of the fix to just show how it works and does not include an update of the settings file or a slightly more sophisticated logic that I used to have different borders for inline and block previews.